// expense-approval-app-vibe-coding/src/generated/models/SharePointModel.ts
export interface SPListExpandedUser {
  // user claims
  Claims?: string;
  // user title
  DisplayName?: string;
  // user email
  Email?: string;
  // user picture
  Picture?: string;
  // user department
  Department?: string;
  // user job title
  JobTitle?: string;
  "@odata.type"?: string;
}

export interface SPListItemAttachment {
  // File identifier
  Id?: string;
  // Link to attachment
  AbsoluteUri?: string;
  // Name
  DisplayName?: string;
}